requirements.txt
openai
python-dotenv
langchain
BeautifulSoup
fastapi
Chroma
python-multipart

